fix: move MeetingDetailFrame to apps page file
